Create README.md
Instructions: There are two files:
1. CreditCardSegmentation.ipynb (main file)
2. userPrediction.ipynb (for single input for users)

After saving the files in one single folder, just need to run the main file and all the output/prediction can be seen the main file only. Even the second file will also be running from the main file. To fit or train a model, our machines take a lot of time because of which I have saved my model with the help of ‘joblib’ so, it will not take much time to run now.